### 2020-11-23 #2853 @hugoledoux
potential reviewers:

  1. clausnagel
  1. ganys
  1. chenkianwee
  1. https://3dom.fbk.eu/people/profile/eozdemir (not sure on github though)
  1. https://research.utwente.nl/en/persons/sander-oude-elberink (not sure on github though)
