### 2020-11-24 #2853 @arfon
:wave: @clausnagel @chenkianwee @ganys - would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission we'd be asking you to review is `3dfier: automatic reconstruction of 3D city models`.

- The software package is here: https://github.com/tudelft3d/3dfier
- The short paper here: [joss.02853/10.21105.joss.02853.pdf](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02853/joss.02853/10.21105.joss.02853.pdf)
