### 2020-11-24 #2853 @arfon
@GANys - excellent, thanks! We generally aim to have JOSS reviews complete within six weeks, but given their iterative nature, it's often most effective if you can start your review in the first couple of weeks (to give initial feedback to the authors).

Does that sound OK?
