### 2020-11-26 #2866 @whedon
[REVIEW]: 3dfier: automatic reconstruction of 3D city models
**Submitting author:** @hugoledoux (<a href="http://orcid.org/0000-0002-1251-8654">Hugo Ledoux</a>)
**Repository:** <a href="https://github.com/tudelft3d/3dfier" target ="_blank">https://github.com/tudelft3d/3dfier</a>
**Version:** v1.3.1
