### 2020-11-26 #2866 @hugoledoux
@arfon the Fig1 in my paper is very small, which is weird because the original .png is rather large (1500px X 346px). I couldn't find a way to make it use the width of the page. Possible?
