### 2020-11-26 #2866 @arfon
@hugoledoux - I think you can force a specific size, e.g. `![](file.jpg){ width=50% }`. Can you try experimenting with syntax like that?
