### 2020-12-02 #2866 @GANys
Just a question about the following point :

_Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)_

Authors claim that "A simple model can be created even using a laptop with less RAM and slower CPU.". My tests confirm the statement but does it correspond to the performance check box requirements? Or should it be a normative benchmark?
