### 2020-12-03 #2866 @GANys
I have finished testing 3Dfier and reading the software paper. I hava some comments. Here is my point-by-point list of comments in regard of the review checklist :

Community Guidelines: besides the common GitHub community standards and tools, I don't see any framework or guidelines for people to contribute.
State of the field: 3Dfier is kind of unique. I don't see any package or solution that provide such an output. Some FME Workbench could for instance perform something comparable but this is completely out of the Open Source community.

Besides the performance related question, I don't have any remaining comment on the checklist.

But, about the paper, I have quick remarks:

- A point is lacking at the end of the sentence _[...] different software_. at the end of the Summary.
- _The resulting surface aims at being error-free._ This sentence does not do justice to the actual outputs of the programme. The main advantage of 3Dfier is that output objects are discrete. "Resulting surface" might means that the output is a common mesh. The result goes far away from it, in a good manner I mean.
- This point is somehow related to the second, but it does not seem obvious that 2D input information should not be contiguous. Hence, the result will not be contiguous in this case but 3Dfier still propose a solution.
It is important to make this clear, all the more so because that the explanation of the "stitch" processing looks to work vertically form Figure 2. Moreover, Figure 2 does not leave a choice on the continuity of the 2D dataset.

Finally, 3Dfier is a state of the art software that improves the creation and thus the usability of consistent 3D City Models. It is used in research since several years and continues to improve.

@hugoledoux Here is an up.
