### 2020-12-08 #2866 @arfon
> Authors claim that "A simple model can be created even using a laptop with less RAM and slower CPU.". My tests confirm the statement but does it correspond to the performance check box requirements? Or should it be a normative benchmark?

Apologies for missing this earlier @GANys. I think you can check this off based on your test here.
