### 2020-12-11 #2866 @chenkianwee
I have issue installing the binary version of 3dfier on my windows machine. This is the error I am getting. Am I suppose to install something before running the windows.zip file?
![image](https://user-images.githubusercontent.com/13062600/101960300-22a14800-3bd5-11eb-8ce3-cb372b5777c1.png)
