### 2020-12-12 #2866 @chenkianwee
I got it fix adter downloading necessary Visual studio packages
