### 2020-12-12 #2866 @chenkianwee
I have followed the example here (http://tudelft3d.github.io/3dfier/first_run.html). I successfully generated the 3D geometry. However, the mesh that was generated are overlapping as shown in the image below. The terrain and side of the buildings have overlapping surfaces, which is not as described in the paper.
![image](https://user-images.githubusercontent.com/13062600/101996402-0cac8980-3ca0-11eb-91bf-3a20b29e5336.png)
