### 2020-12-12 #2866 @chenkianwee
3Dfier is a useful piece of software, especially in the built environment field, where the creation of 3D data is essential for running certain analyses and simulations (e.g. solar, airflow simulation). I see it as a much-needed software that can extend existing 2D GIS tools to generate 3D data for the built environment field.

Community Guidelines: I do not see any guideline or framework for contributing to the project. However, I feel the software is well-managed and there is sufficient documentation for interested parties to extend the software. Considering the software is a standalone cmd tool. It could potentially be a useful module where it could be easily included in other projects for 3Dfying 2D GIS data. So rather than be extended, I think this software is useful for other projects that require 3D generation functionality.

State of the field: There is no mention of the state of the field, there are existing commercial tools that offer similar functionality such as FME and ArcGIS software. They are definitely not in the realm of open-source, but I think it will be useful to mention tools of similar functionality regardless of whether they are proprietary or open-source.

There are still some minor issues with the software, e.g. with the installations and the overlapping surfaces shown in my previous comments. I hope the authors are able to fix those. Once addressed, I recommend the publication of the paper.
