### 2020-12-15 #2866 @hugoledoux
Thanks for comments @GANys, sorry for delay in responding.

> Community Guidelines: besides the common GitHub community standards and tools, I don't see any framework or guidelines for people to contribute.

You are right, I forgot to add this! Will do so very soon.

> State of the field: 3Dfier is kind of unique. I don't see any package or solution that provide such an output. Some FME Workbench could for instance perform something comparable but this is completely out of the Open Source community.

If you see comments below from @chenkianwee he says FME and others could do the same. I somewhat disagree but it's a good point to mention. In short, you can extruse buildings and other other features/objects, but the stitching part that aims obtaining topologically clean data is not possible.

I'll add 1-2 sentence in the paper.
