### 2020-12-15 #2866 @hugoledoux
> I got it fix adter downloading necessary Visual studio packages

Should this be added in the installation tutorial though? Maybe we had this all along and never noticed?
