### 2020-12-15 0af1ef4
Add contributing file, well how to contribute to the project
