+++ CONTRIBUTING.md
@@ -1,0 +1,9 @@
+# Contributing
+
++ The GitHub Issues section is the primary communication channel, please post issues, questions, feature requests there.
+
++ The repository follows the [Gitflow branching model](http://nvie.com/posts/a-successful-git-branching-model/).
+
++ To contribute **documentation** consider extending, fixing or adjusting the `/docs` and submit a pull request.
+
++ To contribute **code**, please submit a pull request.
