### 2020-12-15 1dd6855
Make fig1 wider
