### 2020-12-15 5be90b7
Address comment about the paper from @ganys
