### 2020-12-15 e351e5a
Add commercial software link to paper
as asked by @orenkiwi
