### 2020-12-15 #2866 @hugoledoux
> Community Guidelines: I do not see any guideline or framework for contributing to the project. However, I feel the software is well-managed and there is sufficient documentation for interested parties to extend the software. Considering the software is a standalone cmd tool. It could potentially be a useful module where it could be easily included in other projects for 3Dfying 2D GIS data. So rather than be extended, I think this software is useful for other projects that require 3D generation functionality.
>

I would agree with this, extending it is somewhat cumbersome, but it's possible. Would you say it would be better if I rephrased the paper as such?

> State of the field: There is no mention of the state of the field, there are existing commercial tools that offer similar functionality such as FME and ArcGIS software. They are definitely not in the realm of open-source, but I think it will be useful to mention tools of similar functionality regardless of whether they are proprietary or open-source.

https://github.com/tudelft3d/3dfier/commit/e351e5a20bff6349368a6ea0591890558173eb3f

>
> There are still some minor issues with the software, e.g. with the installations and the overlapping surfaces shown in my previous comments. I hope the authors are able to fix those. Once addressed, I recommend the publication of the paper.

Thanks. Overlap == we aim at avoiding those, and I'd like to know the location then I'll check. We don't guarantee clean 3D output, only if you have clean 2D then can you get sensible 3D output. Garbage in? Garbage out...
