### 2020-12-15 #2866 @chenkianwee
>
>
> > I have followed the example here (http://tudelft3d.github.io/3dfier/first_run.html). I successfully generated the 3D geometry. However, the mesh that was generated are overlapping as shown in the image below. The terrain and side of the buildings have overlapping surfaces, which is not as described in the paper.
> > ![image](https://user-images.githubusercontent.com/13062600/101996402-0cac8980-3ca0-11eb-91bf-3a20b29e5336.png)
>
> Is this in the dataset given in the folder `example_data/*`? If yes, where is it? And more importantly, it's possible that there's a gap there between the terrain and the building, I would be very surprise they overlap?

Yes, this is from the example dataset. There is no overlap in the original dataset, but I think what is happening is that the terrain is inheriting some of the elevations from the building when the terrain meets the buildings. I have uploaded my data here if it is useful for you to see what is wrong?
[example_data.zip](https://github.com/openjournals/joss-reviews/files/5697217/example_data.zip)
