### 2020-12-15 #2866 @chenkianwee
>
>
> > I got it fix adter downloading necessary Visual studio packages
>
> Should this be added in the installation tutorial though? Maybe we had this all along and never noticed?

There is documentation for some other missing visual studio files. Its a different visual studio file. But I notice you have already updated your installation instruction here https://tudelft3d.github.io/3dfier/installation.html
