### 2020-12-15 #2866 @chenkianwee
> > State of the field: 3Dfier is kind of unique. I don't see any package or solution that provide such an output. Some FME Workbench could for instance perform something comparable but this is completely out of the Open Source community.
>
> If you see comments below from @chenkianwee he says FME and others could do the same. I somewhat disagree but it's a good point to mention. In short, you can extruse buildings and other other features/objects, but the stitching part that aims obtaining topologically clean data is not possible.
>
> I'll add 1-2 sentence in the paper
> **update**: done [tudelft3d/3dfier@e351e5a](https://github.com/tudelft3d/3dfier/commit/e351e5a20bff6349368a6ea0591890558173eb3f)

I agree with Hugo that the stitching part is more complex and there might not be an existing function to do that. I overlooked that part when giving the comment about the state of the art. I think it will be beneficial to the paper if maybe you can mention the current state of the art with extrusion etc.
