### 2020-12-17 #2866 @hugoledoux
> Yes, this is from the example dataset.

I'm trying to replicate your issue @chenkianwee , with my macOS, if I compile the master branch, I get this, which is also what the output we get and put in the repository:
![IMAGE 2020-12-17 15:41:16](https://user-images.githubusercontent.com/1546518/102501954-4d304c80-407e-11eb-8569-54b0bbed1b08.jpg)

Now if I look at your .yml file, it's not the same at the one currently in the repo (https://github.com/tudelft3d/3dfier/blob/master/example_data/testarea_config.yml), you removed many parameters that make the spikes appear (I think). Could you run with Windows (compiled by yourself or you used our binary?) with that .yml config file and check if the weird spikes are gone?
