### 2020-12-22 #2866 @chenkianwee
Sorry for the late reply. @hugoledoux I tried with the new yml file and it worked. It seems like I have downloaded an outdated example file here (https://github.com/tudelft3d/3dfier/releases/tag/example_data).  Its on the Github documentation page.
![image](https://user-images.githubusercontent.com/13062600/102935072-6288ea80-4473-11eb-9682-1f4d9c334afe.png). Will be good to update the information to avoid confusion. Thanks.
