### 2021-01-04 #2866 @hugoledoux
hmmm, I wasn't aware of that old-and-weird release. I have just deleted it, since I can imagine it could cause confusion!

The proper and up-to-date `.yml` file is the one in the master branch in the `example_data` folder.
