### 2021-01-19 #2866 @hugoledoux
@arfon seems the service to build pdf (and check DOIs) is down, tried elsewhere too nothing works
