### 2021-01-19 f67b758
Fix typo in paper, thanks @ipadjen
