### 2021-01-19 #2866 @arfon
> @arfon seems the service to build pdf (and check DOIs) is down, tried elsewhere too nothing works

Whedon was being a little sloooooow this morning. Should be fine now.
