+++ docs/pages/getting-started/installation.md
@@ -112,7 +112,8 @@ Download and compile *LAStools* with the following:
 ```
 wget  http://lastools.github.io/download/LAStools.zip
 unzip LAStools.zip
-cd LAStools; mkdir build; cd build
+cd LAStools; mv LASlib/src/LASlib-config.cmake LASlib/src/laslib-config.cmake
+mkdir build; cd build
 cmake ..
 sudo make install
 ```
