### 2021-01-22 10afff1
Initial code prep for binary stl
