+++ src/Building.h
@@ -41,6 +41,7 @@ public:
   void          construct_building_walls(const NodeColumn& nc);
   void          get_obj(std::unordered_map< std::string, unsigned long > &dPts, int lod, std::string mtl, std::string &fs);
   void          get_stl(std::unordered_map< std::string, unsigned long > &dPts, int lod, std::string &fs);
+  void          get_stl_binary(std::unordered_map< std::string, unsigned long > &dPts, int lod, std::string &fs);
   void          get_citygml(std::wostream& of);
   void          get_citygml_imgeo(std::wostream& of);
   void          get_citygml_lod1(std::wostream& of);
