+++ src/Map3d.cpp
@@ -453,6 +453,38 @@ void Map3d::get_stl(std::wostream& of) {
   }
 }
 
+void Map3d::get_stl_binary(std::wostream& of) {
+    std::unordered_map< std::string, unsigned long > dPts;
+    std::string fs[7];
+
+    for (int c = 0; c < 7; c++) {
+        for (auto& p : _lsFeatures) {
+            if (p->get_class() == c) {
+                if (p->get_class() == BUILDING) {
+                    Building* b = dynamic_cast<Building*>(p);
+                    b->get_stl_binary(dPts, _building_lod, fs[0]);
+                }
+                else {
+                    p->get_stl_binary(dPts, fs[c]);
+                }
+            }
+        }
+    }
+
+    // Define each class as a separate solid
+    fs[0] = "solid Building\n"   + fs[0]; fs[0] += "endsolid Building";
+    fs[1] = "solid Water\n"      + fs[1]; fs[1] += "endsolid Water";
+    fs[2] = "solid Bridge\n"     + fs[2]; fs[2] += "endsolid Bridge";
+    fs[3] = "solid Road\n"       + fs[3]; fs[3] += "endsolid Road";
+    fs[4] = "solid Terrain\n"    + fs[4]; fs[4] += "endsolid Terrain";
+    fs[5] = "solid Forest\n"     + fs[5]; fs[5] += "endsolid Forest";
+    fs[6] = "solid Separation\n" + fs[6]; fs[6] += "endsolid Separation";
+
+    for (int c = 0; c < 7; c++) {
+        of << fs[c] << std::endl;
+    }
+}
+
 
 bool Map3d::get_postgis_output(std::string connstr, bool pdok, bool citygml) {
 #if GDAL_VERSION_MAJOR < 2
