+++ docs/_data/sidebars/3dfier_sidebar.yml
@@ -3,7 +3,7 @@
 entries:
 - title: sidebar
   product: 3dfier
-  version: v1.3.0
+  version: v1.3.1
   logo: 3dfierLogo.png
   folders:
   - title:
