### 2021-01-25 #2866 @hugoledoux
@arfon what you asked for:

  - DOI: http://dx.doi.org/10.5281/zenodo.4461324
  - it's for v1.3.1
