### 2021-01-26 #2866 @arfon
@hugoledoux - can you merge this PR? There's an issue with one of your DOIs: https://github.com/tudelft3d/3dfier/pull/124
