### 2021-01-26 #2866 @hugoledoux
done.

I should re-issue/update v1.3.1 now, right?
