### 2021-01-26 #2866 @arfon
> I should re-issue/update v1.3.1 now, right?

No, we don't need an update for something as small as this.
