### 2021-01-26 7b49c85
Added binary STL output
