### 2021-01-26 #2866 @arfon
One more small set of changes @hugoledoux: https://github.com/tudelft3d/3dfier/pull/125
